Move testSetSession, testResetSession to engine-only tests
This also removes global registration of `testing_catalog`.
