+++ testing/trino-tests/src/test/java/io/trino/tests/TestDistributedEngineOnlyQueries.java
@@ -14,7 +14,11 @@
 package io.trino.tests;
 
 import io.trino.Session;
+import io.trino.metadata.Catalog;
+import io.trino.metadata.SessionPropertyManager;
 import io.trino.plugin.memory.MemoryQueryRunner;
+import io.trino.server.testing.TestingTrinoServer;
+import io.trino.testing.DistributedQueryRunner;
 import io.trino.testing.QueryRunner;
 import org.intellij.lang.annotations.Language;
 import org.testng.annotations.Test;
@@ -22,8 +26,10 @@ import org.testng.annotations.Test;
 import java.time.ZonedDateTime;
 import java.util.regex.Pattern;
 
+import static io.airlift.testing.Closeables.closeAllSuppress;
 import static io.trino.SystemSessionProperties.ENABLE_DYNAMIC_FILTERING;
 import static io.trino.sql.analyzer.FeaturesConfig.JoinDistributionType.BROADCAST;
+import static io.trino.testing.TestingSession.createBogusTestingCatalog;
 import static io.trino.testing.sql.TestTable.randomTableSuffix;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
@@ -35,7 +41,32 @@ public class TestDistributedEngineOnlyQueries
     protected QueryRunner createQueryRunner()
             throws Exception
     {
-        return MemoryQueryRunner.createQueryRunner();
+        DistributedQueryRunner queryRunner = MemoryQueryRunner.createQueryRunner();
+        addTestingCatalog(queryRunner);
+        return queryRunner;
+    }
+
+    public static void addTestingCatalog(DistributedQueryRunner queryRunner)
+    {
+        try {
+            for (TestingTrinoServer server : queryRunner.getServers()) {
+                addTestingCatalog(server);
+            }
+        }
+        catch (RuntimeException e) {
+            throw closeAllSuppress(e, queryRunner);
+        }
+    }
+
+    private static void addTestingCatalog(TestingTrinoServer server)
+    {
+        // for testing procedures and session properties
+        Catalog bogusTestingCatalog = createBogusTestingCatalog(TESTING_CATALOG);
+        server.getCatalogManager().registerCatalog(bogusTestingCatalog);
+
+        SessionPropertyManager sessionPropertyManager = server.getMetadata().getSessionPropertyManager();
+        sessionPropertyManager.addSystemSessionProperties(TEST_SYSTEM_PROPERTIES);
+        sessionPropertyManager.addConnectorSessionProperties(bogusTestingCatalog.getConnectorCatalogName(), TEST_CATALOG_PROPERTIES);
     }
 
     /**
